# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Property Portfolio Dashboard API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: properties
    description: Property management
  - name: tenants
    description: Tenant management
  - name: maintenance
    description: Maintenance requests
  - name: refurb
    description: Refurbishment tracking
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /properties:
    get:
      operationId: listProperties
      tags: [properties]
      summary: List all properties
      responses:
        "200":
          description: List of properties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Property"
    post:
      operationId: createProperty
      tags: [properties]
      summary: Create a property
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePropertyInput"
      responses:
        "201":
          description: Created property
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"

  /properties/{id}:
    get:
      operationId: getProperty
      tags: [properties]
      summary: Get a property
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Property
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProperty
      tags: [properties]
      summary: Update a property
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePropertyInput"
      responses:
        "200":
          description: Updated property
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProperty
      tags: [properties]
      summary: Delete a property
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /tenants:
    get:
      operationId: listTenants
      tags: [tenants]
      summary: List all tenants
      responses:
        "200":
          description: List of tenants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tenant"
    post:
      operationId: createTenant
      tags: [tenants]
      summary: Create a tenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTenantInput"
      responses:
        "201":
          description: Created tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"

  /tenants/{id}:
    get:
      operationId: getTenant
      tags: [tenants]
      summary: Get a tenant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateTenant
      tags: [tenants]
      summary: Update a tenant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTenantInput"
      responses:
        "200":
          description: Updated tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTenant
      tags: [tenants]
      summary: Delete a tenant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /maintenance:
    get:
      operationId: listMaintenance
      tags: [maintenance]
      summary: List all maintenance requests
      responses:
        "200":
          description: List of maintenance requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MaintenanceRequest"
    post:
      operationId: createMaintenance
      tags: [maintenance]
      summary: Create a maintenance request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMaintenanceInput"
      responses:
        "201":
          description: Created maintenance request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceRequest"

  /maintenance/{id}:
    put:
      operationId: updateMaintenance
      tags: [maintenance]
      summary: Update a maintenance request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMaintenanceInput"
      responses:
        "200":
          description: Updated maintenance request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceRequest"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteMaintenance
      tags: [maintenance]
      summary: Delete a maintenance request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /refurb:
    get:
      operationId: listRefurb
      tags: [refurb]
      summary: List all refurb projects
      responses:
        "200":
          description: List of refurb projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RefurbProject"
    post:
      operationId: createRefurb
      tags: [refurb]
      summary: Create a refurb project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRefurbInput"
      responses:
        "201":
          description: Created refurb project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefurbProject"

  /refurb/{id}:
    put:
      operationId: updateRefurb
      tags: [refurb]
      summary: Update a refurb project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRefurbInput"
      responses:
        "200":
          description: Updated refurb project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefurbProject"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteRefurb
      tags: [refurb]
      summary: Delete a refurb project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /stats:
    get:
      operationId: getStats
      tags: [properties]
      summary: Get portfolio statistics
      responses:
        "200":
          description: Portfolio stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Property:
      type: object
      properties:
        id:
          type: integer
        address:
          type: string
        propertyType:
          type: string
        bedrooms:
          type: integer
        purchasePrice:
          type: number
        currentValue:
          type: number
        monthlyRent:
          type: number
        monthlyMortgage:
          type: number
        monthlyExpenses:
          type: number
        status:
          type: string
          enum: [occupied, vacant, maintenance]
        purchaseDate:
          type: string
        notes:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - address
        - propertyType
        - bedrooms
        - purchasePrice
        - currentValue
        - monthlyRent
        - monthlyMortgage
        - monthlyExpenses
        - status
        - purchaseDate
        - createdAt
        - updatedAt

    CreatePropertyInput:
      type: object
      properties:
        address:
          type: string
        propertyType:
          type: string
        bedrooms:
          type: integer
        purchasePrice:
          type: number
        currentValue:
          type: number
        monthlyRent:
          type: number
        monthlyMortgage:
          type: number
        monthlyExpenses:
          type: number
        status:
          type: string
          enum: [occupied, vacant, maintenance]
        purchaseDate:
          type: string
        notes:
          type: string
      required:
        - address
        - propertyType
        - bedrooms
        - purchasePrice
        - currentValue
        - monthlyRent
        - monthlyMortgage
        - monthlyExpenses
        - status
        - purchaseDate

    Tenant:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        propertyId:
          type: integer
        propertyAddress:
          type: string
        leaseStart:
          type: string
        leaseEnd:
          type: string
        monthlyRent:
          type: number
        depositPaid:
          type: number
        status:
          type: string
          enum: [active, inactive, pending]
        notes:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - firstName
        - lastName
        - email
        - phone
        - propertyId
        - leaseStart
        - leaseEnd
        - monthlyRent
        - depositPaid
        - status
        - createdAt
        - updatedAt

    CreateTenantInput:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        propertyId:
          type: integer
        leaseStart:
          type: string
        leaseEnd:
          type: string
        monthlyRent:
          type: number
        depositPaid:
          type: number
        status:
          type: string
          enum: [active, inactive, pending]
        notes:
          type: string
      required:
        - firstName
        - lastName
        - email
        - phone
        - propertyId
        - leaseStart
        - leaseEnd
        - monthlyRent
        - depositPaid
        - status

    MaintenanceRequest:
      type: object
      properties:
        id:
          type: integer
        propertyId:
          type: integer
        propertyAddress:
          type: string
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [plumbing, electrical, structural, appliances, general, other]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        status:
          type: string
          enum: [open, in_progress, completed, cancelled]
        estimatedCost:
          type: number
        actualCost:
          type: number
        reportedDate:
          type: string
        completedDate:
          type: string
        notes:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - propertyId
        - title
        - description
        - category
        - priority
        - status
        - estimatedCost
        - reportedDate
        - createdAt
        - updatedAt

    CreateMaintenanceInput:
      type: object
      properties:
        propertyId:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [plumbing, electrical, structural, appliances, general, other]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        status:
          type: string
          enum: [open, in_progress, completed, cancelled]
        estimatedCost:
          type: number
        actualCost:
          type: number
        reportedDate:
          type: string
        completedDate:
          type: string
        notes:
          type: string
      required:
        - propertyId
        - title
        - description
        - category
        - priority
        - status
        - estimatedCost
        - reportedDate

    RefurbProject:
      type: object
      properties:
        id:
          type: integer
        propertyId:
          type: integer
        propertyAddress:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [planned, in_progress, completed, on_hold]
        budget:
          type: number
        actualCost:
          type: number
        startDate:
          type: string
        endDate:
          type: string
        contractor:
          type: string
        notes:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - propertyId
        - title
        - description
        - status
        - budget
        - startDate
        - createdAt
        - updatedAt

    CreateRefurbInput:
      type: object
      properties:
        propertyId:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [planned, in_progress, completed, on_hold]
        budget:
          type: number
        actualCost:
          type: number
        startDate:
          type: string
        endDate:
          type: string
        contractor:
          type: string
        notes:
          type: string
      required:
        - propertyId
        - title
        - description
        - status
        - budget
        - startDate

    PortfolioStats:
      type: object
      properties:
        totalProperties:
          type: integer
        occupiedProperties:
          type: integer
        vacantProperties:
          type: integer
        totalPortfolioValue:
          type: number
        totalPurchaseValue:
          type: number
        totalEquity:
          type: number
        monthlyRentIncome:
          type: number
        monthlyMortgageCosts:
          type: number
        monthlyExpenses:
          type: number
        monthlyCashflow:
          type: number
        annualCashflow:
          type: number
        totalTenants:
          type: integer
        openMaintenanceRequests:
          type: integer
        activeRefurbProjects:
          type: integer
        cashflowByMonth:
          type: array
          items:
            $ref: "#/components/schemas/CashflowDataPoint"
        portfolioValueHistory:
          type: array
          items:
            $ref: "#/components/schemas/ValueDataPoint"
        rentByProperty:
          type: array
          items:
            $ref: "#/components/schemas/PropertyRentData"
      required:
        - totalProperties
        - occupiedProperties
        - vacantProperties
        - totalPortfolioValue
        - totalPurchaseValue
        - totalEquity
        - monthlyRentIncome
        - monthlyMortgageCosts
        - monthlyExpenses
        - monthlyCashflow
        - annualCashflow
        - totalTenants
        - openMaintenanceRequests
        - activeRefurbProjects
        - cashflowByMonth
        - portfolioValueHistory
        - rentByProperty

    CashflowDataPoint:
      type: object
      properties:
        month:
          type: string
        income:
          type: number
        expenses:
          type: number
        cashflow:
          type: number
      required:
        - month
        - income
        - expenses
        - cashflow

    ValueDataPoint:
      type: object
      properties:
        month:
          type: string
        value:
          type: number
        equity:
          type: number
      required:
        - month
        - value
        - equity

    PropertyRentData:
      type: object
      properties:
        address:
          type: string
        rent:
          type: number
      required:
        - address
        - rent