# lib/api-zod/src/generated/types/createPropertyInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Property Portfolio Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { CreatePropertyInputStatus } from "./createPropertyInputStatus";

export interface CreatePropertyInput {
  address: string;
  propertyType: string;
  bedrooms: number;
  purchasePrice: number;
  currentValue: number;
  monthlyRent: number;
  monthlyMortgage: number;
  monthlyExpenses: number;
  status: CreatePropertyInputStatus;
  purchaseDate: string;
  notes?: string;
}